requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email_validator==2.2.0
fastapi==0.115.4
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.12
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain-core==0.3.84
langchain-openai==0.2.5
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.7.33
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.20.0
numpy==2.4.4
openai==1.52.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
plotly==5.24.1
protobuf==5.29.6
psycopg2-binary==2.9.9
pyarrow==24.0.0
pyasn1==0.6.3
pycparser==3.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.41.3
streamlit==1.39.0
tabulate==0.9.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.32.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
